feat: :sparkles: Added Header with Author Information
Completed header styles so that an author and post date can be included on any page. With these changes, a front matter option called 'header' was included to make it optional, and the nav background has been edited to become transparent dynamically based on if the header is included or not.