feat: :sparkles: Large Miscellaneous Update
This commit contains a number of things. A paired shortcode config was added for future use. The home layout was renamed to preview for clarification purposes. The ability to modify deadlines using the data-turnredafter attribute was added. The plane easter egg image was updated to load faster. CSS was added to support buttons and header images in the informational pages. Support for backgrounds and heros were added. The home page was renamed to home.md from index.md to make more sense regarding its function. Defaults for hero and background front matter was added.