refactor: :recycle: Separated Components and Improved Customizability.
Separated content, footer, and header into their own components. Allowed only the date to be shown if an author is included but left blank (empty string). Added an order parameter if the default actions -> content -> testimonials order is not desired (this is an array. Any item included will prioritize it before the others so not everything is required to be included). Also ensured nav styles were appropriate (some issues with text color styles in the mobile menu) and fixed some hero responsiveness issues.