fix: :bug: Fixed TOC Anchor Bug
Previously, the scroll event that placed the TOC in view would interrupt scrolling to a specified anchor. Now, this should be fixed. This commit also includes better styles for the SVG follower path, allowing the caps to be seen and the path to be updated if text wraps or changes in any way. This also includes an update to add the hash to the URL when a link is clicked.