feat: :sparkles: Added Icon Preview Clarification
When previewing, one will typically go back and forth between copying a page and pasting it in a different tab. Unfortunately, moving to a different tab or window removes focus and the previous commit did not help track the last copied page. This update changes the clipboard icon to filled once copied as a better alternative.