feat: :sparkles: Added Build Script and Updated Dom Proposal
The build script may be used in the future to trigger a serverless function that runs the script and returns the built page for the purposes of dynamically building content based on a page change (think WYSIWYG editor or the like). The dom proposal update simply improved some wording issues causing some miscommunication amongst faculty.