feat: :construction: Fixed Issue with Position Sticky
Position Sticky relies on the parent size. By setting it here, it limits it to the height of the screen which was not ideal. However, in production there were line breaks that were unable to be edited that were skewing some sizing. Some CSS was added to target these specific breaks and remove them.